requirements.txt
wget
haversine
geopy
prettytable
colorama
tkintermapview
uuid
pillow
pyinstaller